- Refactored Schedule screen buttons (Mark Complete, Instructions, Start) into VStack with full design system styling - Applied smoother white-to-blue gradient in PrimaryBackgroundView based on feedback - Switched Contacts icon from person.fill to phone.fill - Restored old 'monitor check-in' JSON and removed the new one - Removed unused/archived files
