--- TemplateApplication/Archive/InterestingModules.swift
@@ -1,54 +1,0 @@
-//
-// This source file is part of the Stanford Spezi Template Application open-source project
-//
-// SPDX-FileCopyrightText: 2023 Stanford University
-//
-// SPDX-License-Identifier: MIT
-//
-
-import SpeziOnboarding
-import SwiftUI
-
-
-struct InterestingModules: View {
-    @Environment(OnboardingNavigationPath.self) private var onboardingNavigationPath
-    
-    
-    var body: some View {
-        SequentialOnboardingView(
-            title: "Interesting Modules",
-            subtitle: "INTERESTING_MODULES_SUBTITLE",
-            content: [
-                SequentialOnboardingView.Content(
-                    title: "Onboarding",
-                    description: "INTERESTING_MODULES_AREA1_DESCRIPTION"
-                ),
-                SequentialOnboardingView.Content(
-                    title: "HL7 FHIR",
-                    description: "INTERESTING_MODULES_AREA2_DESCRIPTION"
-                ),
-                SequentialOnboardingView.Content(
-                    title: "Contact",
-                    description: "INTERESTING_MODULES_AREA3_DESCRIPTION"
-                ),
-                SequentialOnboardingView.Content(
-                    title: "HealthKit Data Source",
-                    description: "INTERESTING_MODULES_AREA4_DESCRIPTION"
-                )
-            ],
-            actionText: "Next",
-            action: {
-                onboardingNavigationPath.nextStep()
-            }
-        )
-    }
-}
-
-
-#if DEBUG
-#Preview {
-    OnboardingStack {
-        InterestingModules()
-    }
-}
-#endif
