--- TemplateApplication/Archive/Welcome.swift
@@ -1,63 +1,0 @@
-//
-// This source file is part of the Stanford Spezi Template Application open-source project
-//
-// SPDX-FileCopyrightText: 2023 Stanford University
-//
-// SPDX-License-Identifier: MIT
-//
-
-import SpeziOnboarding
-import SwiftUI
-
-
-struct Welcome: View {
-    @Environment(OnboardingNavigationPath.self) private var onboardingNavigationPath
-    
-    
-    var body: some View {
-        OnboardingView(
-            title: "Spezi Template Application",
-            subtitle: "WELCOME_SUBTITLE",
-            areas: [
-                OnboardingInformationView.Content(
-                    icon: {
-                        Image(systemName: "apps.iphone")
-                            .accessibilityHidden(true)
-                    },
-                    title: "The Spezi Framework",
-                    description: "WELCOME_AREA1_DESCRIPTION"
-                ),
-                OnboardingInformationView.Content(
-                    icon: {
-                        Image(systemName: "shippingbox.fill")
-                            .accessibilityHidden(true)
-                    },
-                    title: "Swift Package Manager",
-                    description: "WELCOME_AREA2_DESCRIPTION"
-                ),
-                OnboardingInformationView.Content(
-                    icon: {
-                        Image(systemName: "list.bullet.clipboard.fill")
-                            .accessibilityHidden(true)
-                    },
-                    title: "Spezi Modules",
-                    description: "WELCOME_AREA3_DESCRIPTION"
-                )
-            ],
-            actionText: "Learn More",
-            action: {
-                onboardingNavigationPath.nextStep()
-            }
-        )
-            .padding(.top, 24)
-    }
-}
-
-
-#if DEBUG
-#Preview {
-    OnboardingStack {
-        Welcome()
-    }
-}
-#endif
