--- TemplateApplication/Resources/Archive/ConsentDocument.md
@@ -1,1 +1,0 @@
-Spezi can render consent documents in the markdown format: This is a *markdown* **example**.
