--- TemplateApplication/Resources/Archive/ConsentDocument.md.license
@@ -1,6 +1,0 @@
-
-This source file is part of the Stanford Spezi Template Application open-source project
-
-SPDX-FileCopyrightText: 2023 Stanford University
-
-SPDX-License-Identifier: MIT
